clean up motion handling and add alternate sig
