+++ zhaquirks/xiaomi/aqara/motion_aq2.py
@@ -1,25 +1,22 @@
 """Xiaomi aqara body sensor."""
-import asyncio
-import logging
 
-from zigpy.zcl.clusters.measurement import IlluminanceMeasurement,\
-    OccupancySensing
+from zigpy.zcl.clusters.measurement import (
+    IlluminanceMeasurement, OccupancySensing
+)
 from zigpy.zcl.clusters.security import IasZone
-from zigpy.zcl.clusters.general import Basic, PowerConfiguration,\
-    Identify, Ota
-from zigpy.quirks import CustomCluster
+from zigpy.zcl.clusters.general import (
+    Basic, PowerConfiguration, Identify, Ota
+)
 from zigpy.profiles import zha
-from zhaquirks.xiaomi import BasicCluster, PowerConfigurationCluster,\
-    TemperatureMeasurementCluster, XiaomiCustomDevice
-from zhaquirks import Bus, LocalDataCluster
+from zhaquirks.xiaomi import (
+    BasicCluster, PowerConfigurationCluster, TemperatureMeasurementCluster,
+    XiaomiCustomDevice
+)
+from zhaquirks import Bus
+from .. import MotionCluster, OccupancyCluster
 
 
 XIAOMI_CLUSTER_ID = 0xFFFF
-_LOGGER = logging.getLogger(__name__)
-OCCUPANCY_STATE = 0
-ZONE_STATE = 0
-ON = 1
-OFF = 0
 
 
 class MotionAQ2(XiaomiCustomDevice):
@@ -30,74 +27,6 @@ class MotionAQ2(XiaomiCustomDevice):
         self.battery_size = 9
         self.motionBus = Bus()
         super().__init__(*args, **kwargs)
-
-    class OccupancyCluster(CustomCluster, OccupancySensing):
-        """Occupancy cluster."""
-
-        cluster_id = OccupancySensing.cluster_id
-
-        def __init__(self, *args, **kwargs):
-            """Init."""
-            super().__init__(*args, **kwargs)
-            self._timer_handle = None
-
-        def _update_attribute(self, attrid, value):
-            super()._update_attribute(attrid, value)
-
-            if attrid == OCCUPANCY_STATE and value == ON:
-                if self._timer_handle:
-                    self._timer_handle.cancel()
-                self.endpoint.device.motionBus.listener_event('motion_event')
-                loop = asyncio.get_event_loop()
-                self._timer_handle = loop.call_later(600, self._turn_off)
-
-        def _turn_off(self):
-            self._timer_handle = None
-            self._update_attribute(OCCUPANCY_STATE, OFF)
-
-    class MotionCluster(LocalDataCluster, IasZone):
-        """Motion cluster."""
-
-        cluster_id = IasZone.cluster_id
-
-        def __init__(self, *args, **kwargs):
-            """Init."""
-            super().__init__(*args, **kwargs)
-            self._timer_handle = None
-            self.endpoint.device.motionBus.add_listener(self)
-
-        def motion_event(self):
-            """Motion event."""
-            super().listener_event(
-                'cluster_command',
-                None,
-                ZONE_STATE,
-                [ON]
-            )
-
-            _LOGGER.debug(
-                "%s - Received motion event message",
-                self.endpoint.device._ieee
-            )
-
-            if self._timer_handle:
-                self._timer_handle.cancel()
-
-            loop = asyncio.get_event_loop()
-            self._timer_handle = loop.call_later(120, self._turn_off)
-
-        def _turn_off(self):
-            _LOGGER.debug(
-                "%s - Resetting motion sensor",
-                self.endpoint.device._ieee
-            )
-            self._timer_handle = None
-            super().listener_event(
-                'cluster_command',
-                None,
-                ZONE_STATE,
-                [OFF]
-            )
 
     signature = {
         #  <SimpleDescriptor endpoint=1 profile=260 device_type=263
