+++ zhaquirks/xiaomi/aqara/motion_aq2b.py
@@ -1,23 +1,19 @@
-"""Xiaomi mija body sensor."""
-import logging
+"""Xiaomi aqara body sensor."""
 
-from zigpy.zcl.clusters.general import (
-    Basic, Identify, Ota, Groups, Scenes, LevelControl, OnOff
-)
+from zigpy.zcl.clusters.measurement import IlluminanceMeasurement,\
+    OccupancySensing
+
+from zigpy.zcl.clusters.general import Basic, Ota
 from zigpy.profiles import zha
-from zhaquirks.xiaomi import (
-    BasicCluster, PowerConfigurationCluster, XiaomiCustomDevice
-)
+from zhaquirks.xiaomi import BasicCluster, XiaomiCustomDevice
 from zhaquirks import Bus
 from .. import MotionCluster, OccupancyCluster
 
-
 XIAOMI_CLUSTER_ID = 0xFFFF
-_LOGGER = logging.getLogger(__name__)
 
 
-class Motion(XiaomiCustomDevice):
-    """Custom device representing mija body sensors."""
+class MotionAQ2(XiaomiCustomDevice):
+    """Custom device representing aqara body sensors."""
 
     def __init__(self, *args, **kwargs):
         """Init."""
@@ -26,29 +22,23 @@ class Motion(XiaomiCustomDevice):
         super().__init__(*args, **kwargs)
 
     signature = {
-        #  <SimpleDescriptor endpoint=1 profile=260 device_type=263
+        #  <SimpleDescriptor endpoint=1 profile=260 device_type=260
         #  device_version=1
-        #  input_clusters=[0, 65535, 3, 25]
-        #  output_clusters=[0, 3, 4, 5, 6, 8, 25]>
+        #  input_clusters=[0, 65535, 1030, 1024]
+        #  output_clusters=[0, 25]>
         1: {
             'manufacturer': 'LUMI',
-            'model': 'lumi.sensor_motion',
+            'model': 'lumi.sensor_motion.aq2',
             'profile_id': zha.PROFILE_ID,
             'device_type': zha.DeviceType.DIMMER_SWITCH,
             'input_clusters': [
                 Basic.cluster_id,
                 XIAOMI_CLUSTER_ID,
-                Ota.cluster_id,
-                Identify.cluster_id
+                OccupancySensing.cluster_id,
+                IlluminanceMeasurement.cluster_id,
             ],
             'output_clusters': [
                 Basic.cluster_id,
-                Ota.cluster_id,
-                Identify.cluster_id,
-                Groups.cluster_id,
-                OnOff.cluster_id,
-                LevelControl.cluster_id,
-                Scenes.cluster_id,
                 Ota.cluster_id
             ],
         },
@@ -58,21 +48,16 @@ class Motion(XiaomiCustomDevice):
         'endpoints': {
             1: {
                 'manufacturer': 'LUMI',
-                'model': 'lumi.sensor_motion',
-                'device_type': zha.DeviceType.OCCUPANCY_SENSOR,
+                'model': 'lumi.sensor_motion.aq2',
                 'input_clusters': [
                     BasicCluster,
-                    PowerConfigurationCluster,
-                    Identify.cluster_id,
+                    IlluminanceMeasurement.cluster_id,
                     OccupancyCluster,
                     MotionCluster,
                     XIAOMI_CLUSTER_ID
                 ],
                 'output_clusters': [
                     Basic.cluster_id,
-                    Identify.cluster_id,
-                    Groups.cluster_id,
-                    Scenes.cluster_id,
                     Ota.cluster_id
                 ],
             }
